# Improve performance of msg-pack