- new
```
[info] Benchmark                                              Mode  Cnt       Score       Error  Units
[info] MessagePackerBenchmark.decodeCirceAST                 thrpt   10   21043.178 ±  5225.806  ops/s
[info] MessagePackerBenchmark.decodeInt10                    thrpt   10  115443.330 ± 23974.484  ops/s
[info] MessagePackerBenchmark.decodeInt30                    thrpt   10   48937.644 ±  8353.555  ops/s
[info] MessagePackerBenchmark.decodeLong10                   thrpt   10  130149.158 ± 15897.320  ops/s
[info] MessagePackerBenchmark.decodeLong30                   thrpt   10   53477.419 ±  9743.527  ops/s
[info] MessagePackerBenchmark.decodeString1000_30            thrpt   10   21026.332 ±   790.009  ops/s
[info] MessagePackerBenchmark.decodeString1000_30_multibyte  thrpt   10   19577.196 ±   404.020  ops/s
[info] MessagePackerBenchmark.decodeString100_10             thrpt   10   62195.970 ±  1343.196  ops/s
[info] MessagePackerBenchmark.decodeString100_30             thrpt   10   20617.921 ±  1536.787  ops/s
```

- old
```
[info] Benchmark                                              Mode  Cnt      Score      Error  Units
[info] MessagePackerBenchmark.decodeCirceAST                 thrpt   10    646.564 ±   14.772  ops/s
[info] MessagePackerBenchmark.decodeInt10                    thrpt   10  17716.005 ±  431.987  ops/s
[info] MessagePackerBenchmark.decodeInt30                    thrpt   10   5111.668 ± 1050.866  ops/s
[info] MessagePackerBenchmark.decodeLong10                   thrpt   10  17483.510 ±  481.919  ops/s
[info] MessagePackerBenchmark.decodeLong30                   thrpt   10   3750.520 ± 2264.415  ops/s
[info] MessagePackerBenchmark.decodeString1000_30            thrpt   10    534.469 ±  116.509  ops/s
[info] MessagePackerBenchmark.decodeString1000_30_multibyte  thrpt   10      9.390 ±    2.339  ops/s
[info] MessagePackerBenchmark.decodeString100_10             thrpt   10   1248.537 ±  676.556  ops/s
[info] MessagePackerBenchmark.decodeString100_30             thrpt   10    445.955 ±  238.035  ops/s

```